3.2!

+++ src/ircd.c
@@ -1001,21 +1001,22 @@ int InitwIRCD(int argc, char *argv[])
 				  portnum = portarg;
 			  break;
 		  case 's':
-			  (void)printf("sizeof(aClient) == %u\n",
-			      sizeof(aClient));
-			  (void)printf("sizeof(aChannel) == %u\n",
-			      sizeof(aChannel));
-			  (void)printf("sizeof(aServer) == %u\n",
-			      sizeof(aServer));
-			  (void)printf("sizeof(Link) == %u\n", sizeof(Link));
-			  (void)printf("sizeof(anUser) == %u\n",
-			      sizeof(anUser));
-			  (void)printf("sizeof(aTKline) == %u\n",
-			      sizeof(aTKline));
-			  (void)printf("sizeof(struct ircstatsx) == %u\n",
-			      sizeof(struct ircstatsx));
-			  (void)printf("aClient remote == %u\n",
-			      CLIENT_REMOTE_SIZE);
+			  (void)printf("sizeof(aClient) == %ld\n",
+			      (long)sizeof(aClient));
+			  (void)printf("sizeof(aChannel) == %ld\n",
+			      (long)sizeof(aChannel));
+			  (void)printf("sizeof(aServer) == %ld\n",
+			      (long)sizeof(aServer));
+			  (void)printf("sizeof(Link) == %ld\n",
+			      (long)sizeof(Link));
+			  (void)printf("sizeof(anUser) == %ld\n",
+			      (long)sizeof(anUser));
+			  (void)printf("sizeof(aTKline) == %ld\n",
+			      (long)sizeof(aTKline));
+			  (void)printf("sizeof(struct ircstatsx) == %ld\n",
+			      (long)sizeof(struct ircstatsx));
+			  (void)printf("aClient remote == %ld\n",
+			      (long)CLIENT_REMOTE_SIZE);
 			  exit(0);
 			  break;
 #ifndef _WIN32
